Move execution context to its own file
