+++ addon/-private/execution_context/acceptance.js
@@ -1,0 +1,17 @@
+export default function AcceptanceExecutionContext(pageObjectNode) {
+  this.pageObjectNode = pageObjectNode;
+}
+
+AcceptanceExecutionContext.prototype = {
+  run(cb) {
+    /* global wait */
+    wait().then(() => {
+      cb(this);
+    });
+  },
+
+  click(selector, container) {
+    /* global click */
+    click(selector, container);
+  }
+};
