+++ addon/-private/execution_context/integration.js
@@ -1,0 +1,28 @@
+import Ember from 'ember';
+
+const { $, run } = Ember;
+
+export default function IntegrationExecutionContext(pageObjectNode, testContext) {
+  this.pageObjectNode = pageObjectNode;
+  this.testContext = testContext;
+}
+
+IntegrationExecutionContext.prototype = {
+  run(cb) {
+    run(() => {
+      cb(this);
+    });
+  },
+
+  click(selector, container) {
+    this.$(selector, container).click();
+  },
+
+  $(selector, container) {
+    if (container) {
+      return $(selector, container);
+    } else {
+      return this.testContext.$(selector);
+    }
+  }
+};
